Generate interface ID in bandwidth usage rate fn, reorganize to new file
